Refactor populate
