+++ examples/getstarted/config/database.js
@@ -4,6 +4,7 @@ const sqlite = {
     filename: '.tmp/data.db',
   },
   useNullAsDefault: true,
+  debug: true,
 };
 
 const postgres = {
