+++ examples/getstarted/src/api/address/controllers/address.js
@@ -2,8 +2,23 @@ const { createCoreController } = require('@strapi/strapi').factories;
 
 module.exports = createCoreController('api::address.address', {
   async find(ctx) {
-    const { results } = await strapi.service('api::address.address').find();
+    // const { results } = await strapi.service('api::address.address').find();
 
-    ctx.body = await this.sanitizeOutput(results);
+    const r = await strapi.db.query('api::address.address').load(
+      {
+        id: 1,
+      },
+      'categories',
+      {
+        limit: 2,
+        orderBy: 'asc|desc',
+      }
+    );
+
+    console.log(r);
+
+    ctx.body = r;
+
+    // ctx.body = await this.sanitizeOutput(results);
   },
 });
