+++ packages/core/database/lib/entity-manager/index.js
@@ -1063,6 +1063,10 @@ const createEntityManager = (db) => {
       return { ...entity, ...entry };
     },
 
+    async LoadPage() {
+
+    },
+
     // TODO: add lifecycle events
     async load(uid, entity, fields, params) {
       const { attributes } = db.metadata.get(uid);
