+++ packages/core/database/lib/query/helpers/populate/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const applyPopulate = require('./apply');
+const processPopulate = require('./process');
+
+module.exports = {
+  applyPopulate,
+  processPopulate,
+};
