+++ packages/core/database/lib/query/query-builder.js
@@ -326,14 +326,14 @@ const createQueryBuilder = (uid, db, initialState = {}) => {
     processSelect() {
       state.select = state.select.map((field) => helpers.toColumnName(meta, field));
 
-      if (this.shouldUseDistinct()) {
-        const joinsOrderByColumns = state.joins.flatMap((join) => {
-          return _.keys(join.orderBy).map((key) => this.aliasColumn(key, join.alias));
-        });
-        const orderByColumns = state.orderBy.map(({ column }) => column);
+      // if (this.shouldUseDistinct()) {
+      //   const joinsOrderByColumns = state.joins.flatMap((join) => {
+      //     return _.keys(join.orderBy).map((key) => this.aliasColumn(key, join.alias));
+      //   });
+      //   const orderByColumns = state.orderBy.map(({ column }) => column);
 
-        state.select = _.uniq([...joinsOrderByColumns, ...orderByColumns, ...state.select]);
-      }
+      //   state.select = _.uniq([...joinsOrderByColumns, ...orderByColumns, ...state.select]);
+      // }
     },
 
     getKnexQuery() {
